+++ common/compositor/renderer.h
@@ -36,7 +36,7 @@ class Renderer {
   virtual bool Draw(const std::vector<RenderState>& commands,
                     NativeSurface* surface) = 0;
 
-  virtual void InsertFence(int kms_fence) = 0;
+  virtual void InsertFence(uint64_t kms_fence) = 0;
 };
 
 }  // namespace hwcomposer
